+++ docs/md/releasenotes.md
@@ -2,6 +2,7 @@
 
 ### Improved
 
+- Allow class sub-hierarchy re-initialization for the optimized implementation of `ClassSubHierarchyInitializer`.
 - Enable providing distinct abort operators for the main and trace reconstruction CPAs.
 - Add a heap model for taint CPA supporting tainting of whole objects.
 - `Call` API: Add a few utilities and fix inconsistent call argument count getter behavior.